Move classify_write to io_utils and document step by example
Two review findings on the pump state machine:

- lib.rs had grown to 402 lines, over the 400-line maximum in AGENTS.md.
  Move classify_write into io_utils, which already owns handle_write,
  WriteOutcome, and the non-fatal error partition it collapses, so the
  write boundary's final step sits with the rest of that boundary. lib.rs
  drops to 388 lines; io_utils/mod.rs (335) and pump_machine.rs (242) stay
  well inside the cap.
- AGENTS.md requires function documentation to demonstrate usage and
  outcome. Add worked examples to pump_machine::step (a completed write,
  then a broken pipe latching closed, draining, and stopping on EOF) and
  to classify_write (complete versus latched-closed after the reader
  hangs up).

Behaviour is unchanged: nextest 59/59 and Kani 5/5 harnesses still pass.
